ScreenKey: Fix thread blocking issue using the self-pipe trick
When closing ScreenKey, the main thread will block on joining with
the listening thread in ViewWidget::~ViewWidget(), while the listening
thread is blocking on reading key events until the user presses some
keys. So the user has to press at least one key to actually terminate
ScreenKey, which is stupid and weird.

By using the self-pipe trick, when it is time to exit, the main thread
will write a byte to the pipe, so that the listening thread which is
blocking on polling the fds set will be notified to wake up and exit
to join with the main thread successfully.